Move `ty::Visibility` to its own little module (cute)
