+++ compiler/rustc_middle/src/ty/mod.rs
@@ -148,7 +148,9 @@ mod structural_impls;
 mod sty;
 mod typeck_results;
 
-// Data types
+mod visibility;
+
+pub use visibility::Visibility;
 
 pub struct ResolverOutputs {
     pub global_ctxt: ResolverGlobalCtxt,
@@ -278,14 +280,6 @@ impl fmt::Display for ImplPolarity {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Copy, Hash, Encodable, Decodable, HashStable)]
-pub enum Visibility<Id = LocalDefId> {
-    /// Visible everywhere (including in other crates).
-    Public,
-    /// Visible only in the given crate-local module.
-    Restricted(Id),
-}
-
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash, HashStable, TyEncodable, TyDecodable)]
 pub enum BoundConstness {
     /// `T: Trait`
@@ -369,57 +363,6 @@ impl TyCtxt<'_> {
         true
     }
 }
-
-impl<Id> Visibility<Id> {
-    pub fn is_public(self) -> bool {
-        matches!(self, Visibility::Public)
-    }
-
-    pub fn map_id<OutId>(self, f: impl FnOnce(Id) -> OutId) -> Visibility<OutId> {
-        match self {
-            Visibility::Public => Visibility::Public,
-            Visibility::Restricted(id) => Visibility::Restricted(f(id)),
-        }
-    }
-}
-
-impl<Id: Into<DefId>> Visibility<Id> {
-    pub fn to_def_id(self) -> Visibility<DefId> {
-        self.map_id(Into::into)
-    }
-
-    /// Returns `true` if an item with this visibility is accessible from the given module.
-    pub fn is_accessible_from(self, module: impl Into<DefId>, tcx: TyCtxt<'_>) -> bool {
-        match self {
-            // Public items are visible everywhere.
-            Visibility::Public => true,
-            Visibility::Restricted(id) => tcx.is_descendant_of(module.into(), id.into()),
-        }
-    }
-
-    /// Returns `true` if this visibility is at least as accessible as the given visibility
-    pub fn is_at_least(self, vis: Visibility<impl Into<DefId>>, tcx: TyCtxt<'_>) -> bool {
-        match vis {
-            Visibility::Public => self.is_public(),
-            Visibility::Restricted(id) => self.is_accessible_from(id, tcx),
-        }
-    }
-}
-
-impl Visibility<DefId> {
-    pub fn expect_local(self) -> Visibility {
-        self.map_id(|id| id.expect_local())
-    }
-
-    /// Returns `true` if this item is visible anywhere in the local crate.
-    pub fn is_visible_locally(self) -> bool {
-        match self {
-            Visibility::Public => true,
-            Visibility::Restricted(def_id) => def_id.is_local(),
-        }
-    }
-}
-
 /// The crate variances map is computed during typeck and contains the
 /// variance of every item in the local crate. You should not use it
 /// directly, because to do so will make your pass dependent on the
